Refactor tests to use primitive types in assertions, replacing boxed values.

+++ src/test/java/dk/tbsalling/aismessages/ais/messages/AddressedBinaryMessageTest.java
@@ -25,10 +25,10 @@ public class AddressedBinaryMessageTest {
         // Assert
         assertEquals(AISMessageType.AddressedBinaryMessage, aisMessage.getMessageType());
         AddressedBinaryMessage message = (AddressedBinaryMessage) aisMessage;
-        assertEquals(Integer.valueOf(0), message.getRepeatIndicator());
+        assertEquals(0, message.getRepeatIndicator());
         assertEquals(MMSI.valueOf(232009504), message.getSourceMmsi());
-        assertEquals((Integer) 0, message.getDesignatedAreaCode());
-        assertEquals((Integer) 29, message.getFunctionalId());
+        assertEquals(0, message.getDesignatedAreaCode());
+        assertEquals(29, message.getFunctionalId());
         // TODO : check the binary value
         assertEquals("0010001111001100000000100111111100000101100101110101101100111100000000000000000100000000000000100000000000001110000011100000000000001110111100100011100000000000000000000111100010010010100101100101101100000000", message.getBinaryData());
     }
@@ -56,7 +56,7 @@ public class AddressedBinaryMessageTest {
         assertInstanceOf(NumberOfPersonsOnBoard.class, asm);
         NumberOfPersonsOnBoard numberOfPersonsOnBoard = (NumberOfPersonsOnBoard) asm;
         assertEquals("0000000000011000", numberOfPersonsOnBoard.getBinaryData());
-        assertEquals(Integer.valueOf(3), numberOfPersonsOnBoard.getNumberOfPersons());
+        assertEquals(3, numberOfPersonsOnBoard.getNumberOfPersons());
    }
 
 }
